removed older files and finalized entire set up
